refactor theme toggle
