+++ app/server.py
@@ -224,7 +224,12 @@ def create_app():
     app.register_blueprint(meme_bp)
 
     @app.route('/')
-    def home(): return render_template('index.html')
+    def home():
+        return render_template('index.html')
+
+    @app.route('/explore')
+    def explore():
+        return render_template('explore.html')
 
     @app.route('/api/stickers')
     def get_stickers_api():
