+++ app/templates/index.html
@@ -38,6 +38,7 @@
         <header class="relative text-center py-6 bg-gray-200 dark:bg-gray-900 border-b border-gray-300 dark:border-gray-700">
             <h1 class="text-3xl md:text-4xl font-bold tracking-tight text-gray-800 dark:text-white">AI Face Swap <span class="text-blue-600">Studio Pro</span></h1>
             <p class="text-md text-gray-500 dark:text-gray-400 mt-1">Workflow Professionale Guidato</p>
+            <a href="{{ url_for('explore') }}" class="absolute left-4 top-4 text-sm text-blue-600 hover:underline">Galleria</a>
             <button id="theme-toggle" class="absolute right-4 top-4 p-2 rounded-md bg-gray-300 dark:bg-gray-700 text-gray-800 dark:text-gray-100 hover:bg-gray-400 dark:hover:bg-gray-600 transition-colors" aria-label="Attiva tema chiaro/scuro">
                 <svg id="theme-icon" class="w-5 h-5" xmlns="http://www.w3.org/2000/svg" fill="currentColor" viewBox="0 0 24 24"></svg>
             </button>
